Add websockets with an useful example
